+++ packages/common/streamsUtils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @moralisweb3/common-streams-utils
 
+## 2.14.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-aptos-utils@2.14.3
+  - @moralisweb3/common-core@2.14.3
+  - @moralisweb3/common-evm-utils@2.14.3
+
+## 2.14.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/common-aptos-utils@2.14.2
+  - @moralisweb3/common-core@2.14.2
+  - @moralisweb3/common-evm-utils@2.14.2
+
 ## 2.14.1
 
 ### Patch Changes